Ensure percent hrs on solar is decimal
It could later be float or whatever, point is it shouldn't be int.

+++ tests/SolarTally.UnitTests/Builders/ApplianceUsageBuilder.cs
@@ -12,7 +12,7 @@ namespace SolarTally.UnitTests.Builders
         public int TestQuantity => 2;
         public decimal TestPowerConsumption => 20.5m;
         public int TestNumHours => 3;
-        public int TestPercentHrsOnSolar => 1;
+        public decimal TestPercentHrsOnSolar => 1;
         public bool TestEnabled => true;
 
         public ApplianceUsageBuilder()
